19-04-2025: CleanUp and changes to remove warnings.

+++ OpenGL/src/XZPlaneGrid.cpp
@@ -29,7 +29,7 @@ void XZPlaneGrid::UpdateAspectRation(float aspectRatio)
 
 void XZPlaneGrid::UpdateGridSize(float gridSize)
 {
-    m_GridSize = gridSize;
+    m_GridSize = static_cast<size_t>(gridSize);
 }
 
 void XZPlaneGrid::UpdateNear(float near)
